// src/Interfaces/Cereal.ts
export interface Cereal {
    id: number;
    name: string;
    manufacturer: string;
    type: string;
    calories: number;
    protein: number;
    fat: number;
    sodium: number;
    fiber: number;
    carbohydrates: number;
    sugars: number;
    potassium: number;
    vitamins: number;
    shelf: number;
    weight: number;
    cups: number;
    rating: string;
}